finished adding last pages

+++ home.py
@@ -2,6 +2,7 @@ from flask import Flask, render_template,request, redirect, request
 import webbrowser
 from threading import Timer
 # import cozmo_random_behaviors
+import randomBehavior
 import json
 import os
 import csv
@@ -14,9 +15,9 @@ def home():
 
 @app.route('/startRandomBehavior', methods=['POST'])
 def startRandomBehavior(): 
-    print("it ran")
     # cozmo_random_behaviors.start()
-    return "hello"
+    randomBehavior.start()
+    return "success"
 
 def open_browser():
       webbrowser.open_new('http://127.0.0.1:8080/')
@@ -27,9 +28,11 @@ def enterDataInFile():
     data = request.form["form"]
     data = data.split("&")
     dataAsDict = {}
+
     for key in data:
         key = key.split("=")
         dataAsDict[key[0]] = key[1]
+
     interest_alarm = dataAsDict["likert2"]
     confusion_understanding = dataAsDict["likert3"]
     frusteration_relief = dataAsDict["likert4"]
@@ -38,17 +41,28 @@ def enterDataInFile():
     fear_hope = dataAsDict["likert7"]
     boredom_surprise = dataAsDict["likert8"]
     disgust_desire = dataAsDict["likert9"]
-    try:
-        with open("data.csv", 'a') as file:
+    
+    if os.path.isfile("data.csv"):
+         with open("data.csv", 'a') as file:
             writer = csv.writer(file)
-            writer.writerows([description, interest_alarm, confusion_understanding, frusteration_relief, sorrow_joy, anger_gratitude, fear_hope, boredom_surprise, disgust_desire])
-    except IOError:
+            writer.writerow([description, interest_alarm, confusion_understanding, frusteration_relief, sorrow_joy, anger_gratitude, fear_hope, boredom_surprise, disgust_desire])
+    else:
         with open("data.csv", 'w') as file:
             writer = csv.writer(file)
             writer.writerow(["Description", "Interst/Alarm", "Confusion/Understanding", "Frusteration/Relief","Sorrow/Joy", "Anger/Gratitude","Fear/Hope", "Boredom/Surprise","Disgust/Desire"])
-            writer.writerow([description, interest_alarm, confusion_understanding, frusteration_relief, sorrow_joy, anger_gratitude, fear_hope, boredom_surprise, disgust_desire])
-       
+            writer.writerow([description, interest_alarm, confusion_understanding, frusteration_relief, sorrow_joy, anger_gratitude, fear_hope, boredom_surprise, disgust_desire])    
+    return "success"
 
+@app.route('/addEmail', methods=["POST"])
+def addEmail():
+    userEmail = request.form["email"]
+    if os.path.isfile("email.txt"):
+        with open("email.txt", 'a') as file:
+            file.write("\n" + userEmail)
+    else:
+        with open("email.txt", 'w') as file:
+            file.write(userEmail)
+    return "success"
 
 if __name__ == '__main__':
     Timer(1, open_browser).start()
